fix 9-sum_total.py - bugg

+++ math/0x02-calculus/9-sum_total.py
@@ -6,4 +6,4 @@ def summation_i_squared(n):
     '''sumatory of i squared'''
     if (type(n) != int and type(n) != float) or n == 0:
         return None
-    return sum(range(1, (n * n) + 1, n))
+    return int(sum(range(1, (n * n) + 1, n)))
